Refactor langs module: separate execution logic into executor module
- Split langs module into minimal language definition module and new executor module
- langs module now only handles language definitions (langs.toml parsing and Lang struct)
- executor module contains all Docker execution functionality (compile, execute, monitoring)
- Moved sources directory and embedded test files to executor module
- Improved separation of concerns for better modularity
