+++ executor/sources/aplusb/ac_func.cpp
