+++ executor/sources/aplusb/cpp-func/fastio.h
