+++ executor/sources/aplusb/cpp-func/solve.hpp
