+++ executor/sources/aplusb/sample.in
