+++ executor/sources/badcode/use_many_stack.cpp
